refactor: create modes namespace
